apps/script/requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
googleapis-common-protos==1.72.0
grpcio==1.76.0
idna==3.11
importlib_metadata==8.7.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
protobuf==6.33.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
zipp==3.23.0
